+++ packages/pn-commons/src/services/tracking.service.ts
@@ -27,6 +27,26 @@ export function trackEvent(event_name: string, nodeEnv: string, properties?: any
   }
 }
 
+/**
+ * Set profile properties
+ */
+export function setProfileProperty(property: any, nodeEnv: string): void {
+  if (!nodeEnv || nodeEnv === 'development') {
+    // eslint-disable-next-line no-console
+    console.log('Mixpanel events mock on console log - profile properties');
+  } else if (nodeEnv === 'test') {
+    return;
+  } else {
+    try {
+      mixpanel.identify(mixpanel.get_distinct_id());
+      mixpanel.people.set(property);
+    } catch (_) {
+      // eslint-disable-next-line no-console
+      console.log(property);
+    }
+  }
+}
+
 export const interceptDispatch =
   (
     next: Dispatch<AnyAction>,