+++ packages/pn-commons/src/utility/__test__/notification.utility.test.ts
@@ -7,6 +7,7 @@ import {
   notificationDTOMultiRecipient,
   notificationToFe,
   notificationToFeMultiRecipient,
+  notificationToFeRadd,
   payments,
 } from '../../__mocks__/NotificationDetail.mock';
 import {
@@ -1040,6 +1041,23 @@ describe('parse notification & filters', () => {
         (viewedElement.details as ViewedDetails).delegateInfo!.taxId
       })`
     );
+  });
+
+  it('return radd timeline element when radd is present in timeline', () => {
+    const raddElement = {
+      elementId: 'NOTIFICATION_RADD_RETRIEVED_mock',
+      timestamp: '2022-06-21T11:44:28Z',
+      legalFactsIds: [],
+      category: TimelineCategory.NOTIFICATION_RADD_RETRIEVED,
+      details: {
+        recIndex: 1,
+        eventTimestamp: '2022-06-21T11:44:28Z',
+        raddType: 'ALT',
+        raddTransactionId: '6',
+      },
+    };
+    const raddFromTimeline = notificationToFeRadd.radd;
+    expect(raddElement).toStrictEqual(raddFromTimeline);
   });
 });