+++ packages/pn-pa-webapp/src/pages/NotificationDetail.page.tsx
@@ -3,7 +3,7 @@ import React, { Fragment, useCallback, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router-dom';
 
-import { Alert, Box, Grid, Paper, Stack, Typography } from '@mui/material';
+import { Alert, AlertTitle, Box, Grid, Paper, Stack, Typography } from '@mui/material';
 import {
   ApiError,
   AppResponse,
@@ -16,7 +16,6 @@ import {
   NotificationDetail as NotificationDetailType,
   NotificationRelatedDowntimes,
   PnBreadcrumb,
-  TimedMessage,
   TitleBox,
   appStateActions,
   dateIsLessThan10Years,
@@ -95,12 +94,6 @@ const NotificationDetail: React.FC = () => {
   const legalFactDownloadUrl = useAppSelector(
     (state: RootState) => state.notificationState.legalFactDownloadUrl
   );
-  const legalFactDownloadRetryAfter = useAppSelector(
-    (state: RootState) => state.notificationState.legalFactDownloadRetryAfter
-  );
-  const legalFactDownloadAARRetryAfter = useAppSelector(
-    (state: RootState) => state.notificationState.legalFactDownloadAARRetryAfter
-  );
 
   const { recipients } = notification;
   /*
@@ -123,7 +116,20 @@ const NotificationDetail: React.FC = () => {
     if (_.isObject(document)) {
       void dispatch(
         getSentNotificationOtherDocument({ iun: notification.iun, otherDocument: document })
-      );
+      )
+        .unwrap()
+        .then((response) => {
+          if (response.retryAfter) {
+            dispatch(
+              appStateActions.addInfo({
+                title: '',
+                message: t(`detail.document-not-available`, {
+                  ns: 'notifiche',
+                }),
+              })
+            );
+          }
+        });
     } else {
       const documentIndex = document as string;
       void dispatch(getSentNotificationDocument({ iun: notification.iun, documentIndex }));
@@ -145,7 +151,20 @@ const NotificationDetail: React.FC = () => {
             category: legalFactAsLegalFact.category,
           },
         })
-      );
+      )
+        .unwrap()
+        .then((response) => {
+          if (response.retryAfter) {
+            dispatch(
+              appStateActions.addInfo({
+                title: '',
+                message: t(`detail.document-not-available`, {
+                  ns: 'notifiche',
+                }),
+              })
+            );
+          }
+        });
     } else if ((legalFact as NotificationDetailOtherDocument).documentId) {
       const otherDocument = legalFact as NotificationDetailOtherDocument;
       void dispatch(getSentNotificationOtherDocument({ iun: notification.iun, otherDocument }));
@@ -245,9 +264,6 @@ const NotificationDetail: React.FC = () => {
   useDownloadDocument({ url: documentDownloadUrl });
   useDownloadDocument({ url: otherDocumentDownloadUrl });
 
-  const timeoutMessage = legalFactDownloadRetryAfter * 1000;
-  const timeoutAARMessage = legalFactDownloadAARRetryAfter * 1000;
-
   const properBreadcrumb = (
     <PnBreadcrumb
       linkRoute={routes.DASHBOARD}
@@ -314,13 +330,20 @@ const NotificationDetail: React.FC = () => {
                     downloadFilesMessage={getDownloadFilesMessage('attachments')}
                     downloadFilesLink={t('detail.download-files-link', { ns: 'notifiche' })}
                   />
+                  {notification.radd && (
+                    <Alert severity={'success'} sx={{ mb: 3, mt: 2 }} data-testid="raddAlert">
+                      <AlertTitle>
+                        {t('detail.timeline.radd.title', { ns: 'notifiche' })}
+                      </AlertTitle>
+                      {notification.recipients.length === 1
+                        ? t('detail.timeline.radd.description-mono-recipient', { ns: 'notifiche' })
+                        : t('detail.timeline.radd.description-multi-recipients', {
+                            ns: 'notifiche',
+                          })}
+                    </Alert>
+                  )}
                 </Paper>
                 <Paper sx={{ p: 3, mb: 3 }} elevation={0} data-testid="aarDownload">
-                  <TimedMessage timeout={timeoutAARMessage}>
-                    <Alert severity={'warning'} sx={{ mb: 3 }} data-testid="aarNotAvailableAlert">
-                      {t('detail.document-not-available', { ns: 'notifiche' })}
-                    </Alert>
-                  </TimedMessage>
                   <NotificationDetailDocuments
                     title={t('detail.aar-acts', { ns: 'notifiche' })}
                     documents={notification.otherDocuments ?? []}
@@ -343,11 +366,6 @@ const NotificationDetail: React.FC = () => {
             </Grid>
             <Grid item lg={5} xs={12}>
               <Box sx={{ backgroundColor: 'white', height: '100%', p: 3, pb: { xs: 0, lg: 3 } }}>
-                <TimedMessage timeout={timeoutMessage}>
-                  <Alert severity={'warning'} sx={{ mb: 3 }} data-testid="docNotAvailableAlert">
-                    {t('detail.document-not-available', { ns: 'notifiche' })}
-                  </Alert>
-                </TimedMessage>
                 <NotificationDetailTimeline
                   language={i18n.language}
                   recipients={recipients}