+++ packages/pn-personafisica-login/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.2](https://github.com/pagopa/pn-frontend/compare/v2.3.1...v2.3.2) (2024-02-20)
+
+
+### Bug Fixes
+
+* **PN-10025:** added aria-label to button with IDP name ([#1141](https://github.com/pagopa/pn-frontend/issues/1141)) ([2554c60](https://github.com/pagopa/pn-frontend/commit/2554c60c302dab1ebb8784040fc0c8fc2a225e5f))
+
+
+
+
+
+## [2.3.1](https://github.com/pagopa/pn-frontend/compare/v2.3.1-RC.1...v2.3.1) (2024-02-19)
+
+**Note:** Version bump only for package @pagopa-pn/pn-personafisica-login
+
+
+
+
+
+## [2.3.1-RC.0](https://github.com/pagopa/pn-frontend/compare/v2.3.0...v2.3.1-RC.0) (2024-02-14)
+
+**Note:** Version bump only for package @pagopa-pn/pn-personafisica-login
+
+
+
+
+
 # [2.3.0](https://github.com/pagopa/pn-frontend/compare/v2.3.0-RC.2...v2.3.0) (2024-01-25)
 
 **Note:** Version bump only for package @pagopa-pn/pn-personafisica-login